requirements.txt
fastapi
uvicorn
openai
faiss-cpu
numpy
python-multipart
pypdf2
python-docx
pandas
tiktoken